+++ android/src/main/java/ly/img/react_native/pesdk/RNPhotoEditorSDKModule.kt
@@ -34,7 +34,8 @@ import java.io.File
 class RNPhotoEditorSDKModule(val reactContext: ReactApplicationContext) : ReactContextBaseJavaModule(reactContext), ActivityEventListener, PermissionListener {
 
     companion object {
-        const val EDITOR_RESULT_ID = 1
+        // This number must be unique. It is public to allow client code to change it if the same value is used elsewhere.
+        var EDITOR_RESULT_ID = 489313435
     }
 
     init {
@@ -52,7 +53,8 @@ class RNPhotoEditorSDKModule(val reactContext: ReactApplicationContext) : ReactC
         IMGLY.authorize()
     }
 
-    override fun onActivityResult(activity: Activity, requestCode: Int, resultCode: Int, data: Intent) {
+    override fun onActivityResult(activity: Activity, requestCode: Int, resultCode: Int, resultData: Intent?) {
+        val data = resultData ?: return // If resultData is null the result is not from us.
         when (requestCode) {
             EDITOR_RESULT_ID -> {
                 when (resultCode) {